backend/python-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
pydantic==2.11.0
pydantic-settings==2.7.0
python-multipart==0.0.15
httpx==0.28.1
aiohttp==3.10.11
python-json-logger==2.0.7
structlog==24.4.0
rich==13.9.4
colorlog==6.8.2
tenacity==9.0.0
# RAG Embedding Dependencies
sentence-transformers==3.3.1
numpy==2.2.1
torch>=2.6.0
transformers==4.47.1
# Data Ingestion Dependencies
yt-dlp==2024.12.23
youtube-transcript-api==1.2.2
PyPDF2==3.0.1
python-docx==1.2.0
ebooklib==0.19
beautifulsoup4==4.12.2
readability-lxml==0.8.4.1
requests==2.32.3
aiofiles==23.2.1
nltk==3.9.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
# Note: Run 'python -m spacy download en_core_web_sm' after installation
